save list to local storage
